Add Hardware Security Module wrapper script interface
Instead of specifying (-k) with the path of a private key the
name of hardware security wrapper script may be provided.

The script must accept the filename of the data to sign and emit
a PKCS#1 v1.5 signature encoded as hex to stdout.

There is no facility for passing arbitrary arguments to the wrapper
script. It is assumed that it specific to the key e.g. it could
be written to tmpfs (mktemp) before invoking this update-pieeprom.sh